requirements.txt
aiogram>=3.5.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0